requirements.txt
torch
pytorch-lightning
torchmetrics
fair-esm
jsonargparse
wandb
tokenizers
transformers
beartype
datasets
transformers[torch]
sphinx
sphinx-book-theme
